+++ .babelrc
@@ -1,0 +1,7 @@
+{
+  "env": {
+    "development": {
+      "plugins": ["transform-es2015-modules-commonjs"]
+    }
+  }
+}
