+++ .eslintrc.js
@@ -1,3 +1,6 @@
 module.exports = {
-    "extends": "airbnb"
+  "extends": "airbnb",
+  "env": {
+    "jest": true,
+  }
 };
