+++ __tests__/sum.js
@@ -1,0 +1,3 @@
+module.exports = function sum(a, b) {
+  return a + b;
+};
