+++ package.json
@@ -4,9 +4,9 @@
   "description": "> Project description",
   "main": ".eslintrc.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
     "start": "nodemon server.js",
-    "react-dev": "webpack -d --watch"
+    "react-dev": "webpack -d --watch",
+    "test": "./node_modules/.bin/jest"
   },
   "repository": {
     "type": "git",
@@ -25,6 +25,8 @@
     "@babel/preset-react": "^7.0.0",
     "babel": "^6.23.0",
     "babel-loader": "^8.0.2",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-react": "^6.24.1",
     "css-loader": "^1.0.0",
     "eslint": "^5.5.0",
     "eslint-config-airbnb": "^17.1.0",
@@ -45,6 +47,12 @@
     "request": "^2.88.0",
     "style-loader": "^0.23.0",
     "webpack": "^4.17.2",
-    "webpack-cli": "^3.1.0"
+    "webpack-cli": "^3.1.0",
+    "yarn": "^1.9.4"
+  },
+  "standard": {
+    "env": [
+      "jest"
+    ]
   }
 }
